+++ src/main/java/com/meoguri/linkocean/domain/profile/service/command/FavoriteService.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.service;
+package com.meoguri.linkocean.domain.profile.service.command;
 
 public interface FavoriteService {
