+++ src/main/java/com/meoguri/linkocean/domain/profile/service/query/CategoryQueryService.java
@@ -1,10 +1,10 @@
-package com.meoguri.linkocean.domain.bookmark.service;
+package com.meoguri.linkocean.domain.profile.service.query;
 
 import java.util.List;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 
-public interface CategoryService {
+public interface CategoryQueryService {
 
 	/* 사용자가 작성한 북마크가 존재하는 카테고리 목록 조회 */
 	List<Category> getUsedCategories(long profileId);
